chore: move omnichannel to one single folder
