+++ packages/core-typings/src/index.ts
@@ -105,7 +105,7 @@ export * from './IOmnichannelCannedResponse';
 export * from './ILivechatMonitorRecord';
 export * from './ILivechatDepartment';
 export * from './IOmnichannelAgent';
-export * from './OmichannelRoutingConfig';
+export * from './omnichannel/OmichannelRoutingConfig';
 export * from './IVoipExtension';
 export * from './voip';
 export * from './ACDQueues';
@@ -113,7 +113,7 @@ export * from './IVoipConnectorResult';
 export * from './IVoipServerConfig';
 export * from './IVoipServerConnectivityStatus';
 export * from './IOmnichannelVoipServiceResult';
-export * from './IInquiry';
+export * from './omnichannel/IInquiry';
 export * from './ILivechatPriority';
 export * from './ILogs';
 export * from './IOmnichannelServiceLevelAgreements';
