+++ packages/core-typings/src/omnichannel/ILivechatAgentActivity.ts
@@ -1,4 +1,4 @@
-import type { IRocketChatRecord } from './IRocketChatRecord';
+import type { IRocketChatRecord } from '../IRocketChatRecord';
 
 export interface ILivechatAgentActivity extends IRocketChatRecord {
 	agentId: string;
