+++ packages/core-typings/src/omnichannel/ILivechatBusinessHour.ts
@@ -1,4 +1,4 @@
-import type { ILivechatDepartment } from './ILivechatDepartment';
+import type { ILivechatDepartment } from '../ILivechatDepartment';
 
 export enum LivechatBusinessHourTypes {
 	DEFAULT = 'default',
