+++ packages/core-typings/src/omnichannel/ILivechatCustomField.ts
@@ -1,4 +1,4 @@
-import type { IRocketChatRecord } from './IRocketChatRecord';
+import type { IRocketChatRecord } from '../IRocketChatRecord';
 
 export interface ILivechatCustomField extends IRocketChatRecord {
 	label: string;
