+++ packages/core-typings/src/omnichannel/ILivechatDepartment.ts
