+++ packages/core-typings/src/omnichannel/ILivechatDepartmentAgents.ts
@@ -1,4 +1,4 @@
-import type { IRocketChatRecord } from './IRocketChatRecord';
+import type { IRocketChatRecord } from '../IRocketChatRecord';
 
 export interface ILivechatDepartmentAgents extends IRocketChatRecord {
 	departmentId: string;
