+++ packages/core-typings/src/omnichannel/ILivechatDepartmentRecord.ts
@@ -1,4 +1,4 @@
-import type { IRocketChatRecord } from './IRocketChatRecord';
+import type { IRocketChatRecord } from '../IRocketChatRecord';
 
 /** @deprecated */
 export interface ILivechatDepartmentRecord extends IRocketChatRecord {
