+++ packages/core-typings/src/omnichannel/ILivechatMonitorRecord.ts
@@ -1,4 +1,4 @@
-import type { IRocketChatRecord } from './IRocketChatRecord';
+import type { IRocketChatRecord } from '../IRocketChatRecord';
 
 export interface ILivechatMonitorRecord extends IRocketChatRecord {
 	_id: string;
