+++ packages/core-typings/src/omnichannel/IOmnichannel.ts
