+++ packages/core-typings/src/omnichannel/IOmnichannelBusinessUnit.ts
@@ -1,4 +1,4 @@
-import type { IRocketChatRecord } from './IRocketChatRecord';
+import type { IRocketChatRecord } from '../IRocketChatRecord';
 
 export interface IOmnichannelBusinessUnit extends IRocketChatRecord {
 	_id: string;
