+++ packages/core-typings/src/omnichannel/IOmnichannelCustomAgent.ts
