+++ packages/core-typings/src/omnichannel/IOmnichannelServiceLevelAgreements.ts
@@ -1,4 +1,4 @@
-import type { IRocketChatRecord } from './IRocketChatRecord';
+import type { IRocketChatRecord } from '../IRocketChatRecord';
 
 export const DEFAULT_SLA_CONFIG = {
 	ESTIMATED_WAITING_TIME_QUEUE: 9999999,
